refactor: FixDiscountController, RateDiscountController 생성
